WinApp/requirements.txt
matplotlib==2.2.3
numpy==1.21.6
onnxruntime==1.11.1
opencv_python==4.6.0.66
Pillow==9.2.0
PyQt5==5.15.7
torch==1.12.0
torchvision==0.13.0
